==> notebooks/0 - Charging Points Check.ipynb <==
import json
import os
import sys

import branca
import branca.colormap as cm
import geopandas as gpd
import numpy as np
import pandas as pd
import json 

sys.path.insert(0, "../src")
from BrusselsMap import BrusselsMap
from GeoJsonHandler import GeoJsonHandler
from OSMGeoJsonHandler import OSMGeoJsonHandler

pd.set_option("display.max_rows", 500)
pd.set_option("display.max_columns", 500)

import warnings
warnings.filterwarnings('ignore')

cars = pd.read_csv("../data/ChargingPointsCheck/TF_CAR_HH_SECTOR.csv", encoding = "utf-8")
pop = pd.read_csv("../data/ChargingPointsCheck/population_stats_sectors.csv", encoding = "utf-8")

cars = cars[['CD_STAT_SECTOR', 'TX_REFNIS_DESCR_FR', 'MS_NUM_CAR']]
cars.head()

pop = pop[['CD_SECTOR', 'TX_DESCR_FR', 'POPULATION']]
pop.rename(columns={'CD_SECTOR':'CD_STAT_SECTOR', 'TX_DESCR_FR':'TX_REFNIS_DESCR_FR'}, inplace=True)
pop.head()

pd.merge(pop, cars, on=['CD_STAT_SECTOR','TX_REFNIS_DESCR_FR'], how='left')

with open('../data/ChargingPointsCheck/RBC_stat_sectors_2021.json') as f:
    data = json.load(f)

gdf = gpd.GeoDataFrame.from_features(data["features"])
gdf.head()

sectors = gdf[['Sector', 'T_SEC_FR']]
sectors.rename(columns={'Sector':'CD_STAT_SECTOR'}, inplace=True)
sectors.head()

len(sectors)

dff= pd.merge(sectors, cars, on='CD_STAT_SECTOR', how='left')
dff.fillna(0, inplace=True)
dff['MS_NUM_CAR'].astype(int)
dff.head()

dff['MS_NUM_CAR'].sum()

# Nombre de garages privés
96053

324243-96053
