==> notebooks/0 - Utils_FromBelgianToGPS.ipynb <==
`Author: Victor Radermecker | Date: 11/10/2021 | Project: Master Thesis`

This notebook allows to convert a shapefile in Belgian National Geographic System to Longitude/Latitude.

## Imports

import branca.colormap as cm
import folium
import geopandas as gpd
import pandas as pd
import pyproj
import shapely.ops as sp_ops
from pyproj import Transformer
from tqdm.auto import tqdm

tqdm.pandas()
pd.set_option("display.max_columns", 100)

## Global Functions and variables

belgian_coords = 31370
gps_coords = 4326

def transform_geom_to_srid(geom, scrs, tcrs):
    """ Transform a geometry to a new target CRS.
        Works with pyproj version >= 2.x.x
        - geom is a Shapely geometry instance
        - scrs is your input CRS EPSG integer (the one of your original data)
        - tcrs is your target CRS EPSG integer (the one you want to reproject
                  your data in, probably 4326 in your case)
    """
    project = Transformer.from_crs(
        "EPSG:" + str(scrs), "EPSG:" + str(tcrs), always_xy=True
    )
    return sp_ops.transform(project.transform, geom)

## Statistical sectors - Preprocessing

# Statistical Sectors
fname = (
    "../data/maps/stat_sectors/shapefile/sh_statbel_statistical_sectors_20210101.shp"
)

# Read and display
ss = gpd.read_file(fname)

# Keep only statistical sectors from RBC and rename 'CS01012021' column to 'Sector'
ss = ss[ss["T_REGIO_FR"] == "Région de Bruxelles-Capitale"]
ss = ss[["CS01012021", "T_REGIO_FR", "T_SEC_FR", "geometry"]]
ss.rename(columns={"CS01012021": "Sector"}, inplace=True)
print("Number of statistical sectors for RBC: ", len(ss))

ss.head()

#### Start conversion 

This cell starts the conversion from the Belgian coordinate system to the universal GPS system.

<div class="alert alert-block alert-warning">
Warning: Takes about an hour (for whole Belgium). RBC is faster.
</div>

ss["geometry"] = ss["geometry"].progress_apply(
    transform_geom_to_srid, args=(belgian_coords, gps_coords,)
)

save_json = True
save_shp = False

# Save to GeoJson file for choropleth maps
if save_json:
    ss.to_file(
        "../data/maps/stat_sectors/json/RBC_stat_sectors_2021.json", driver="GeoJSON"
    )

if save_shp:
    ss.to_file("../data/maps/stat_sectors/shapefile/RBC_stats_sectors_2021_gps.shp")

## Proximus Neighborhoods - Preprocessing

# Statistical Sectors
fname = "../data/maps/prox_neighbor/shapefile/UrbAdm_MONITORING_DISTRICT.shp"

# Read and display
pn = gpd.read_file(fname)

pn.head()

#### Start conversion 

This cell starts the conversion from the Belgian coordinate system to the universal GPS system.

pn["geometry"] = pn["geometry"].progress_apply(
    transform_geom_to_srid, args=(belgian_coords, gps_coords,)
)

save_json = True
save_shp = False

# Save to GeoJson file for choropleth maps
if save_json:
    pn.to_file(
        "../data/maps/prox_neighbor/json/RBC_Neighborhoods_gps.json", driver="GeoJSON"
    )

if save_shp:
    pn.to_file("../data/maps/prox_neighbor/shapefile/RBC_Neighborhoods_gps.json")
