==> notebooks/0 - Utils_GenerateRandomDataset.ipynb <==
`Author: Victor Radermecker | Date: 11/10/2021 | Project: Master Thesis`

This Notebook generates the Proximus Hourly Dataset.

import pandas as pd
import numpy as np
from tqdm import tqdm

colnames = ['NAME_FRE', 'REGULAR', 'IRREGULAR']

prox = pd.read_csv("../data/proximus/proximusFrequentTrip_RandomlyGenerated.csv")
prox.drop(columns=['Unnamed: 0'])
prox.head()

# Generate the random dataset

ngh_names = prox['originNB'].unique()
colnames = ['originNB', 'destinationNB', 'arrival_hr', 'regularTripSample', 'irregularTripSample']
data= pd.DataFrame(columns=colnames)
data.head()

for i in tqdm(ngh_names):
    temp_ = pd.DataFrame()
    temp_['originNB'] = [i for l in range(24*len(ngh_names))]
    temp_['destinationNB'] = np.repeat(ngh_names, 24) 
    temp_['arrival_hr'] = list(np.arange(0,24))*len(ngh_names)
    temp_['regularTripSample'] = pd.Series(np.random.randint(low=0, high=10000, size=24*len(ngh_names)))
    temp_['irregularTripSample'] = pd.Series(np.random.randint(low=0, high=10000, size=24*len(ngh_names)))
    
    data = data.append(temp_)

#Save dataset
data.to_csv('../data/proximus/hourly_proximusFrequentTrip_RandomlyGenerated.csv', index=False)
