==> notebooks/1 - ProximusDataset_AggregatedDemand.ipynb <==
`Author: Victor Radermecker | Date: 11/10/2021 | Project: Master Thesis`

Aggregating the Proximus Dataset for each neighborhood.

## Imports

import os
import sys

import branca.colormap as cm
import folium
import geopandas as gpd
import ipywidgets as widgets
import pandas as pd
from tqdm.auto import tqdm

sys.path.insert(0, "../src")
from BrusselsMap import BrusselsMap
from GeoJsonHandler import GeoJsonHandler

tqdm.pandas()
import json

pd.set_option("display.max_columns", 100)

%load_ext autoreload
%autoreload 2

## Data aggregation for daily proximus dataset

<div class="alert alert-block alert-warning">
Warning!
</div>

`CHAUSSEE DE WAVRE - SAINT-JULIEN ` needs to be replaced by `CHAUSSEE DE WAVRE - SAINT-JULIEN`


`SAINTE-ALIX  - JOLI BOIS` needs to be replaced by `SAINTE-ALIX - JOLI BOIS`

# Create a GeoJson Handler
r_path = r"../data/maps/prox_neighbor/json/RBC_Neighborhoods_gps.json"
path = os.path.join(os.getcwd(), r_path)
sectors = GeoJsonHandler(path, "NAME_FRE")

prox = pd.read_csv(
    "../data/proximus/proximusFrequentTrip_RandomlyGenerated.csv",
    encoding="unicode_escape",
).drop(columns='Unnamed: 0')

#Aggregate data
prox = prox.groupby(by=['destinationNB'])[['regularTripSample','irregularTripSample']].sum().reset_index()
prox.rename(columns={'destinationNB':'NAME_FRE','regularTripSample':'REGULAR','irregularTripSample':'IRREGULAR'}, inplace=True)

prox.head()

prox.to_csv('../data/proximus/proximus_aggregated.csv', index=False)

## Data aggregation for hourly proximus dataset
