==> notebooks/1- Proximus - Antennas.ipynb <==
import json
import os
import sys

import branca
import branca.colormap as cm
import geopandas as gpd
import numpy as np
import pandas as pd

sys.path.insert(0, "../src")
from BrusselsMap import BrusselsMap
from GeoJsonHandler import GeoJsonHandler
from OSMGeoJsonHandler import OSMGeoJsonHandler

pd.set_option("display.max_rows", 500)
pd.set_option("display.max_columns", 500)

import warnings
warnings.filterwarnings('ignore')

import geopandas as gpd
import folium
import matplotlib.pyplot as plt

tacs = gpd.read_file("C:/Users/vic_t/OneDrive/MFE/Code/data/proximus/TACS/cell_locations_20200120_20200220.shp")

# Use WGS 84 (epsg:4326) as the geographic coordinate system
tacs = tacs.to_crs(epsg=4326)

gpd.GeoSeries(tacs.iloc[0]['geometry'])

m = folium.Map(location=[50.85, 4.35], zoom_start=12, tiles='CartoDB positron')
folium.TileLayer("stamenwatercolor").add_to(m)

# style_function = lambda x: {
#     "weight": 0.5,
#     "color": "black",
#     "fillColor": colormap(x['layer'][colnames[1]]),
#     "fillOpacity": 0.65,
# }

fill_style_function = lambda x: {
    "weight": 0.25,
    "color": "black",
    "fillColor": "grey",
    "fillOpacity": 0.35,
}

for _, r in tacs.iterrows():
    # Without simplifying the representation of each borough,
    # the map might not be displayed
    sim_geo = gpd.GeoSeries(r['geometry'])#.simplify(tolerance=0.001)
    geo_j = sim_geo.to_json()
    
#     #Set json
#     a = gpd.GeoSeries(tacs.iloc[i][['geometry']]).to_json()
#     b = json.loads(a)
#     b["features"][i]["properties"] = {'layer':tacs.iloc[i]['layer']}
    
    geo_j = folium.GeoJson(data=geo_j,
                           style_function=fill_style_function)
    folium.Popup(r['id']).add_to(geo_j)
    geo_j.add_to(m)


sim_geo = gpd.GeoSeries(df.iloc[115]['geometry'])#.simplify(tolerance=0.001)
geo_j = sim_geo.to_json()
geo_j = folium.GeoJson(data=geo_j,
                       style_function=fill_style_function)
geo_j.add_to(m)

folium.GeoJson(gpd.GeoSeries(df["geometry"].iloc[115]).to_json()).add_to(m)
    
    
m

o_path = r"../data/outputs/maps/"
m.save(os.path.join(o_path, "Proximus-TACS.html")) # to HTML

len(tacs)
