==> notebooks/6 - Segmentation - Choropleth Map.ipynb <==
import geopandas as gpd

import json
import os
import sys


import branca
import branca.colormap as cm
import numpy as np
import pandas as pd

sys.path.insert(0, "../src")
from BrusselsMap import BrusselsMap
from GeoJsonHandler import GeoJsonHandler
from OSMGeoJsonHandler import OSMGeoJsonHandler

pd.set_option("display.max_rows", 500)
pd.set_option("display.max_columns", 500)

import warnings
warnings.filterwarnings('ignore')

%load_ext autoreload
%autoreload 2

## Segmentation maps
In this notebook, we will plot a choropleth map of two neighborhoods to show two different charging habits

health = pd.read_csv('../data/final_data/osm/health.csv')
culture = pd.read_csv('../data/final_data/osm/culture.csv')
education = pd.read_csv('../data/final_data/osm/school_university.csv')
shop = pd.read_csv('../data/final_data/osm/shop.csv')
sustenance = pd.read_csv('../data/final_data/osm/sustenance.csv')
tourism = pd.read_csv('../data/final_data/osm/tourism.csv')
office = pd.read_csv('../data/final_data/osm/office.csv')

from shapely import wkt

dataframes = [health, culture, education, shop, sustenance, tourism, office]
for df in dataframes:
    df['geometry'] = df['geometry'].apply(wkt.loads)
    df = gpd.GeoDataFrame(df, crs='epsg:4326')

#Load geodata
r_path = r"../data/final_data/sector.json"
path = os.path.join(os.getcwd(), r_path)
sectors = GeoJsonHandler(path, "NAME_FRE")

mymap = BrusselsMap(sectors, 800, 800, tiles=False, tile="stamenwatercolor")

#Adding hospitals
data = health[health['NAME_FRE']== 'VOGELENZANG - ERASME']
data_json = data[['NAME_FRE','AMENITY','geometry']].to_json()
mymap.build_sectors(data_json)

#Adding hospitals
data = health[health['NAME_FRE']== 'VOGELENZANG - ERASME']
data_json = data[['NAME_FRE','AMENITY','geometry']].to_json()
mymap.build_sectors(data_json)

mymap.map

data = health[health['NAME_FRE']== 'VOGELENZANG - ERASME']

for _, r in data.iterrows():
    # Without simplifying the representation of each borough,
    # the map might not be displayed
    # sim_geo = gpd.GeoSeries(r['geometry']).simplify(tolerance=0.001)
    geo_j = r['geometry'].to_json()
    geo_j = folium.GeoJson(data=geo_j,
                           style_function=lambda x: {'fillColor': 'orange'})
    folium.Popup(r['NAME_FRE']).add_to(geo_j)
    geo_j.add_to(mymap.map)
mymap.map

r['geometry']
